Add helpers folder and move createCandleData to it
