+++ src/components/container/MarketDetails.js
@@ -1,13 +1,14 @@
 import React, { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import Chart from 'react-apexcharts';
+import createCandleData from '../../helpers/createCandleData';
 
 const MarketDetails = () => {
   const { search } = useLocation();
   const marketTag = search.split('=')[1];
 
   const API_KEY = 'f3e7a6c62566f5327da3a9f3fd13890e';
-  // const marketTag = 'EURUSD';
+
   const api = `https://financialmodelingprep.com/api/v3/fx/${marketTag}?apikey=${API_KEY}`;
   const historicalApi = `https://financialmodelingprep.com/api/v3/historical-chart/5min/${marketTag}?apikey=${API_KEY}`;
 
@@ -28,19 +29,6 @@ const MarketDetails = () => {
         enabled: true,
       },
     },
-  };
-
-  const createCandleData = (data) => {
-    const candles = data.map((entry) => {
-      const {
-        open, high, low, close, date,
-      } = entry;
-      return ({
-        x: date,
-        y: [open, high, low, close],
-      });
-    });
-    return candles.slice(-300);
   };
 
   const [marketDetails, setMarketDetails] = useState({});
@@ -68,8 +56,9 @@ const MarketDetails = () => {
         const response = await fetch(historicalApi);
         const data = await response.json();
         const createdCandleData = createCandleData(data);
-        console.log(createdCandleData);
-        setCandleData(createdCandleData);
+        if (createdCandleData !== []) {
+          setCandleData(createdCandleData);
+        }
       } catch (error) {
         setCandleDataError(error);
       }
@@ -81,46 +70,56 @@ const MarketDetails = () => {
     ticker, bid, ask, changes, high, low,
   } = marketDetails;
 
-  console.log([marketDetailsError, candleDataError]);
-
   return (
     <section className="item details container mt-3 ">
       <div className="d-flex justify-content-center my-5">
         <div className="col-12 col-md-6">
-          <article className="d-flex flex-column bg-dark text-white p-2">
-            <h5>{ticker}</h5>
-            <div className=" d-flex flex-column">
-              <p className="d-flex justify-content-between">
-                <span>
-                  Bid:
-                  {bid}
-                </span>
-                <span>
-                  Ask:
-                  {ask}
-                </span>
-              </p>
-              <p className="d-flex justify-content-center">
-                arrow
-                {changes}
-              </p>
-              <p className="d-flex justify-content-between">
-                <span>
-                  High:
-                  {high}
-                </span>
-                <span>
-                  Low:
-                  {low}
-                </span>
-              </p>
-            </div>
-          </article>
+          {
+            marketDetailsError === ''
+              ? (
+                <article className="d-flex flex-column bg-dark text-white p-2">
+                  <h5>{ticker}</h5>
+                  <div className=" d-flex flex-column">
+                    <p className="d-flex justify-content-between">
+                      <span>
+                        Bid:
+                        {bid}
+                      </span>
+                      <span>
+                        Ask:
+                        {ask}
+                      </span>
+                    </p>
+                    <p className="d-flex justify-content-center">
+                      arrow
+                      {changes}
+                    </p>
+                    <p className="d-flex justify-content-between">
+                      <span>
+                        High:
+                        {high}
+                      </span>
+                      <span>
+                        Low:
+                        {low}
+                      </span>
+                    </p>
+                  </div>
+                </article>
+              )
+              : <div />
+          }
         </div>
       </div>
-      <div id="chart" className="row">
-        <Chart options={options} series={[{ data: candleData }]} type="candlestick" height={350} />
-      </div>
+      {
+        candleData.length > 0 && candleDataError === ''
+          ? (
+            <div id="chart" className="row">
+              <Chart options={options} series={[{ data: candleData }]} type="candlestick" height={350} />
+            </div>
+          )
+          : <div />
+      }
     </section>
   );
 };
